refactor(worker): split CardanoIndexWorker into partial-class files by concern
Behavior-preserving reorganization of the 1223-line CardanoIndexWorker.cs into
six files of the same partial class. No logic changes: method bodies were moved
byte-identical, and the full method-signature set is verified identical to HEAD
(nothing dropped, renamed, or duplicated).

  CardanoIndexWorker.cs                 ctor + fields + core loop (ExecuteAsync,
                                        StartReducerChainSyncAsync, Exit, Dispose)
  CardanoIndexWorker.DependencyGraph.cs BuildDependencyGraph / BuildPipelines / ...
  CardanoIndexWorker.StartPoints.cs     reducer-state init + dependent start-point adjust
  CardanoIndexWorker.Dashboard.cs       TUI rendering + progress/colour helpers
  CardanoIndexWorker.Telemetry.cs       per-reducer timing aggregation
  CardanoIndexWorker.Logging.cs         the 27 [LoggerMessage] partials

The class was already partial (for the source-generated loggers); this leans on
that to make a hard-to-navigate file reviewable without threading shared state
through new types. True collaborator extraction stays a possible follow-up.
Full solution build clean (0 warnings); 35 EndToEnd tests green.
